+++ sdk/cognitivelanguage/azure-ai-language-conversations/tests/test_conversation_app.py
@@ -63,3 +63,51 @@ class ConversationAppTests(ConversationTest):
             assert result["result"]["prediction"]["entities"][0]["category"] == 'Contact'
             assert result["result"]["prediction"]["entities"][0]["text"] == 'Carol'
             assert result["result"]["prediction"]["entities"][0]["confidenceScore"] > 0
+
+    @pytest.mark.live_test_only
+    @GlobalConversationAccountPreparer()
+    def test_conversation_app_aad_auth(self, endpoint, key, conv_project_name, conv_deployment_name):
+        token = self.get_credential(ConversationAnalysisClient)
+        client = ConversationAnalysisClient(endpoint, token)
+        with client:
+            query = "Send an email to Carol about the tomorrow's demo"
+            result = client.analyze_conversation(
+                task={
+                    "kind": "Conversation",
+                    "analysisInput": {
+                        "conversationItem": {
+                            "participantId": "1",
+                            "id": "1",
+                            "modality": "text",
+                            "language": "en",
+                            "text": query
+                        },
+                        "isLoggingEnabled": False
+                    },
+                    "parameters": {
+                        "projectName": conv_project_name,
+                        "deploymentName": conv_deployment_name,
+                        "verbose": True
+                    }
+                }
+            )
+
+            # assert - main object
+            assert not result is None
+            assert result["kind"] == "ConversationResult"
+
+            # assert - prediction type
+            assert result["result"]["query"] == query
+            assert result["result"]["prediction"]["projectKind"] == 'Conversation'
+
+            # assert - top intent
+            assert result["result"]["prediction"]["topIntent"] == 'Setup'
+            assert len(result["result"]["prediction"]["intents"]) > 0
+            assert result["result"]["prediction"]["intents"][0]["category"] == 'Setup'
+            assert result["result"]["prediction"]["intents"][0]["confidenceScore"] > 0
+
+            # assert - entities
+            assert len(result["result"]["prediction"]["entities"]) > 0
+            assert result["result"]["prediction"]["entities"][0]["category"] == 'Contact'
+            assert result["result"]["prediction"]["entities"][0]["text"] == 'Carol'
+            assert result["result"]["prediction"]["entities"][0]["confidenceScore"] > 0