+++ sdk/cognitivelanguage/azure-ai-language-conversations/azure/ai/language/conversations/_patch.py
@@ -58,6 +58,9 @@ class ConversationAnalysisClient(GeneratedConversationAnalysisClient): # pylint:
     the best candidate service to handle the request. At last, it returns a response with the
     candidate service's response as a payload.
 
+    See https://docs.microsoft.com/rest/api/language/conversation-analysis-runtime/ for more information about
+    requests and responses you can pass to this client.
+
     :param endpoint: Supported Cognitive Services endpoint (e.g.,
      https://:code:`<resource-name>`.cognitiveservices.azure.com). Required.
     :type endpoint: str